requirements.txt
python-telegram-bot==21.6
openai==1.54.0
yt-dlp==2024.10.22
python-dotenv==1.0.1
